Add files via upload
This is the .tft file for the Digital Bird Nextion display. To install this on a Nextion first download and add this file and this file only to a small Mico SD card
Insert the SD card into the Nextion's Micro SD slot before powering up. Power up the system and the Nextion will find the file and install the software.
remove the SD card and restart the display, you should see the Digital Bird user interface. All Done.