Add files via upload
To install on Nextion display first upload the file to a blank micro SD card. Place the card in the Nextion SD Card slot and power up the display. The system should automatically find the file and upload the Digital bird User Interface. This should only take a few seconds. When upload is complete remove the SD card and power down the diaplay. Power the display back up and the digital bird user interface should appear on the screen.